backend/requirements.txt
fastapi
uvicorn[standard]
httpx
shapely
python-geohash
geohash2
python-dotenv
openai
